Make `Label` own its contents
We expect to pass these around after the initial buffer they were parsed
out of (most likely JSON from GitHub) has been deallocated. That
requires us to own the data.

If this becomes a performance concern we can investigate interning the
labels - we are essentially guaranteed to have a relatively small set of
labels around in the long term.