Add nominate command
This allows for the bot to know which labels to add/remove during
nomination and beta backport acceptance, rather than trying to push that
knowledge into all teams. It also enforces adding a team label when
nominating, to aid in folks seeing nominates.

Currently the nominated and beta-nominated/accepted labels are hardcoded
but team name to team label mappings are not. This may want to change to
even less hard coding in the future.