Move team-data API access to separate module
This also adds retries (up to three total requests) if we timeout on connection.
It's possible we should be less restrictive and always retry, but the hope is
that timeouts are sufficient. This may not cover (for example) a "connection
reset by peer" error though, even if fundamentally that's also a timeout, just
of a different kind.