Move database client creation to db module
Also, this moves the download for the AWS TLS certificate to a one-off per run
of triagebot, placing it in a lazy_static. This slightly speeds up subsequent
`make_client()` calls and should work out fine, as in practice updates to that
certificate happen very rarely (and if it does, we'd just restart the app and
get a new one).