Add some debug logging for request building
It's suspected that there's a failure of some kind arising when we attempt to
set labels on GitHub issues, but it isn't yet clear what that failure is.

Our logs so far just say "builder error: unsupported value" which is pretty
unhelpful, though it does seem like it is coming from reqwest (the "builder
error" string at least).