Support multiple different bot names
Currently the triagebot name on GitHub isn't owned by the Rust org, but seems to
be largely unused too. However, we're calling the tool triagebot, and users can
easily get it wrong by calling the tool triagebot on GitHub rather than rustbot
(the "real" name).

For now, just accept both names interchangeably. We do not issue a warning for
the time being.