Fix wrong endpoint for searching in pull requests
The Github API are a bit confusing when the intent is to search
either through issues *or* pull requests. The `/{repo}/pulls` endpoint does
not support filter by label, while the `/{repo}/issues` does.

By the way, under the hood Github only has "issues" as entities and
pull requests are just "issues with a special flag".
https://docs.github.com/en/rest/reference/issues#list-repository-issues

Commit 106bbb22 unearthed this bug.

This patch should fix when searching through pull requests in the
rust-lang/rfcs repository, labels `T-compiler` and `I-nominated` were
ignored, corresponding to the `[RFC]` section of the WG-Prioritization agenda

Also removed the old redundant function to search through pulls.