Add/remove labels instead of setting a new set of labels
Setting the full set of labels means that we may race with additions or removals
by the user, which is not the desired behavior. For example, when
autolabeling triggered by files present in a new PR, adding labels would (at
least sometimes) kill labels added by the user, as those are added "after" the
open event is sent, even if immediately specified by the user.