Bail out if adding a non-existent label
This prevents creating typo'd labels through rustbot, as the GitHub add label
API will implicitly create a new label if asked to add it to an issue.

We also no longer attempt to remove labels if we failed to add labels, though we
can still add labels and then fail to remove them.