Add agenda endpoints to triagebot
This lets users easily fetch up-to-date agendas from triagebot

Currently this endpoint is pretty seriously rate-limited, but we can switch
that to serving a stale copy instead (refreshing in the background) if this is a
problem in practice. I suspect the usage will be rare enough that it won't be a
problem.