Shortcut can terminate on just a word
In theory, this could introduce conflicts with some other command (causing an
error due to two parses succeeding, I believe), but in practice the risk is low
and the current behavior sufficiently annoying that this seems worth it.

It's also a decision we can mostly go back on in the future.