Fix the pull_request event to be marked as a pull_request.
The `pull_request` event was not marking the issue as a pull_request.
That means that some handlers were ignoring pull requests when they
shouldn't have (like the rfc rendered handler).

Additionally I removed the empty `PullRequestDetails` struct and replaced
it with a bool instead. I don't foresee the `PullRequestDetails` ever
getting fields added to it (the fields usually are part of `Issue`).