Adjust RFC rendered linkifier.
This changes the URL used in the RFC "Rendered" link to link directly to
the author's branch instead of a specific commit. The problem with
specific commits is that they need to be updated on every push. The
existing code didn't handle that (it ignored if it already had a link),
and I think SHA blob links are less desirable since they have a risk
that reviewers will be looking at outdated content.

This also drops the check for `Synchronize` events, since this isn't
updating the rendered link on every push.