Improve filter for T-compiler PRs s-waiting-for-review
By adding the "T-compiler" filter straight into the GraphQL query
(instead of filtering them post-query) it seems we are getting a better
"ranking" for this label - some really old PRs are emerging.

Side effect: by building the GraphQL query with 2 labels instead of one,
I get a 502 error when asking for 100 items per page. Reducing
to 50 or 10 items per page and running multiploe queries seems to
workaround the issue.