Rework Activity in how it differentiates a JSON vs non-JSON response.
Not all API responses are JSON, which can make it a bit awkward to
differentiate which kind of responses are JSON. This unifies the
two Activity variants so that there is only one, which simplifies
things a bit. Non-JSON responses are stored as a JSON string, under
the assumption that GitHub never response with a JSON string.