Add new PR review assignment
This is the implementation of a new workflow for assigning pull requests
to the Rust project contributors.

This new workflow will assign a pull request to be reviewed based on the
number of assigned pull requests to team members.

Everytime a pull request assignment is invoked with `r? <team>` or `r?
<team_member>` the Triagebot will check the current workload of the
candidates and assign the pull request to the team member less busy.

The new workflow is DISABLED by default. It can be enabled by setting
the env variable `USE_NEW_PR_ASSIGNMENT` (any value) and restarting the
Triagebot.

Teams that are subject to the new PR review assignment are listed as a
comma separated list in the env variable `NEW_PR_ASSIGNMENT_TEAMS`. See
`.env.sample` for a usage example.

Team members workload is tracked in a new DB table `review_capacity`.

Both the initial population and the synchronization of this table are
handled by a command line tool that will speak to the Triagebot through
a number of HTTP endpoints. These HTTP endpoints must be PRIVATE and
accessible ONLY to the sync tool. This should be handled at the
infrastructure level, the endpoints are not authenticated.