Set comment created_at/updated_at types to support null.
The `review` field of `PullRequestReviewEvent` isn't quite the same as a
`Comment`, even though we are pretending they are the same. The
`submitted_at` is optional, which was causing deserialization errors.

I was inclined to stop treating pull request review comments the same as
comments, since their fields don't quite line up. However, that is a
much larger change, and generally makes handling pr review comments more
awkward.