Fix cargo milestoning timeout.
This fixes an issue where not all cargo PRs were getting milestoned.

The process of milestoning all the cargo PRs can take more than 10
seconds. This was causing the thread to get cancelled due to the webhook
timeout.

This fixes the problem by spawning a background task to handle the
milestoning process.