Disable GitHub retry on the server.
This disables the GitHub rate limit retry logic on the server. Generally
for webhooks, the server must respond within 10 seconds. If it ever
hits the rate limit, it will likely sleep for several minutes, meaning
it will be guaranteed to be killed.

Whether or not retries are enabled is now a property of `GithubClient`.
It is enabled for the CLI prioritization tool (which added this retry
logic in the first place in
https://github.com/rust-lang/triagebot/pull/579).