Print warning when assigning someone on vacation
People on vacation will usually not (they better >:() respond during
their vacation, but they probably will be able to respond after. There
is no reason why you wouldn't be able to assign to someone on vacation
manually if this person is the best reviewer and you have enough
patience for them to return, so don't prevent it. Do print a warning
though to set expectations for both the author and triage.