Remove label grammar docs
These docs are slightly wrong in various ways (like it doesn't handle
space-separated labels, and `\S` isn't correct since it doesn't allow
certain punctuation).

Instead of trying to keep both this and the user documentation
up-to-date and correct, I'd like to just drop this and lean on the user
documentation.