Improve range-diff color contrast in dark mode
The current colors are very low contrast, and therefore very hard to
read. This replaces the colors with significantly brighter ones to
improve contrast a lot.

The red drifted a lot into the orange, because when making the red
brighter, it just got *harder* to read for me because red on black is
bad combination. So i just made it slightly orange, which looks a lot
better. I did not pick the colors from `git range-diff` itself because
the red there is still pretty low contrast.

The colors now have a comfortable ~4.55 (AA accessibility) contrast
according to developer tools.