docs_update: Skip if a PR is already open
This fixes an issue where triagebot would get stuck in a loop trying to
update a PR every few hours if a PR is already open for a docs update.

The solution here is to just skip the update if a pull request is
already open. This uses a GraphQL search query because it is relatively
cheap (costs 1 graphql point), and there wasn't anything that already
did anything similar.